+++ inst/app/global.R
@@ -11,3 +11,6 @@ addResourcePath("figures_design", "tools/help/figures/")
 ## loading urls and ui
 source("init.R", encoding = getOption("radiant.encoding"), local = TRUE)
 options(radiant.url.patterns = make_url_patterns())
+
+## print options
+options(max.print = max(getOption("max.print"), 5000))